requirements.txt
numpy==1.20.3
opencv-python==4.5.1.48
pandas==1.2.4
pygame==1.9.4
python-dateutil==2.8.1
pytz==2021.1
RPi.GPIO==0.7.0
six==1.16.0
tflite-runtime==2.5.0
